Rename tables.rs -> storage.rs
The main reason is that we have non-table storage types and helpers as
well, and we want a name that ensures we put them all in the storage
affecting places.

Discussed in https://svixhq.slack.com/archives/C0A72988962/p1775850953459599
