+++ crates/auth-token/src/lib.rs
@@ -5,7 +5,7 @@ use fjall_utils::Databases;
 pub mod controller;
 pub mod entities;
 pub mod operations;
-pub mod tables;
+pub mod storage;
 
 use crate::controller::AuthTokenController;
