+++ crates/diom-namespace/src/storage.rs
